fix: 🐛 lerna "packages affected" line, initial state
Fixes an issue where the "packages affected" line that's supposed to get
added to the commit (in the case that the repo is valid lerna repo), is
broken. Root cause was line 20 of runInteractiveQuestions.js, where we
iterate over the initial state key "answers" and only copy over those
that are defined in the initial state - since "packages" was missing
(they key "lerna" was being used instead), the lerna info was getting
lost.