fix: 🐛 fixes scope's empty string answer issue
When you choose an empty string as an answer the answer continues to
live and still listens to user's input while the next question is
started. This causes questions interference and possible answers
overriding. So I cloned the autocomplete-prompt package, fixed this
issue and published to NPM just to fix this particular issue.